Renamed components to something better
