+++ client/src/components/collections/expanded-collection/expanded-collection.style.css
