+++ client/src/components/page-not-found/page-not-found.component.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import './PageNotFound.css';
-
 const PageNotFound = () => (
   <div className="page-not-found">Page not found.</div>
 );
