+++ client/src/components/with-auth/with-auth.component.jsx
