+++ client/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import App from './App';
+import App from './components/app/app.component';
 
 render(<App />, document.querySelector('#app'));
