+++ server/package.json
@@ -16,6 +16,7 @@
     "color": "^3.1.2",
     "concurrently": "^5.2.0",
     "cookie-parser": "^1.4.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.18",
