+++ server/server.js
@@ -1,4 +1,5 @@
 const path = require('path');
+require('dotenv').config()
 
 const express = require('express');
 const mongoose = require('mongoose');
@@ -9,6 +10,9 @@ const verifyToken = require('./utils/verifyToken');
 
 const app = express();
 const PORT = process.env.PORT || 5000;
+
+
+console.log(process.env.DB)
 
 // ------
 // Routes
